pyproject.toml
[project]
name = "srbusapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "dusaan-m", email = "dusanmilivojevic01@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "hishel>=0.1.1",
    "pycryptodome>=3.21.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
